CorrData

CorrData - Objects for ambient noise cross-correlations.

Noise.CorrData — Type.

CorrData

A structure for cross-correlations of ambient noise data.

Fields: CorrData

Field	Description

:name	Freeform channel names

:id	Channel ids. use NET.STA.LOC.CHAN format when possible.

:loc	Location (position) object.

:fs	Sampling frequency in Hz.

:gain	Scalar gain; divide data by the gain to convert to units

:freqmin	Minimum frequency for whitening.

:freqmax	Maximum frequency for whitening.

:cc_len	Length of each correlation in seconds.

:cc_step	Spacing between correlation windows in seconds.

:whitened	Whitening applied.

:time_norm	Apply one-bit whitening with "one_bit".

:resp	Instrument response object, format [zeros poles]

:misc	Dictionary for non-critical information.

:notes	Timestamped notes; includes automatically-logged acquisition and

processing information.

:maxlag	Maximum lag time in seconds to keep from correlations.

:t	Starttime of each correlation.

:corr	Correlations stored in columns.

source