InputParams

Input Parameters - Objects for scheduling cross-correlations

Noise.InputParams — Type.

InputParams

A structure for inputing parameters necessary for processing ambient noise data.

Fields: InputParams

Field	Description

:pair	Network.Station.Channel for each station to be correlated.

:FFTOUT	Directory filepath for FFT output.

:CORROUT	Directory filepath for correlation output.

:cc_len	Length of each correlation in seconds.

:cc_step	Spacing between correlation windows in seconds.

:fs	Sampling frequency in Hz to downsample raw data; set to 0.0.

:freqmin	Minimum frequency for whitening.

:freqmax	Maximum frequency for whitening.

:time_norm	Apply one-bit whitening with "one_bit".

:to_whiten	Apply whitening to FFTs.

:maxlag	Maximum lag time in seconds to keep from correlations.

:smoothinghalfwin	Half window in samples for FFT smoothing.

:corr_type	Type of correlation e.g. "cross-correlate", "deconv" or "coherence"

source