Computing FFTs

Computing FFTs - methods for computing FFTs raw noise data.

Noise.whiten — Function.

whiten(A, freqmin, freqmax, fs, pad=100)

Whiten spectrum of time series A between frequencies freqmin and freqmax. Uses real fft to speed up computation. Returns the whitened (single-sided) fft of the time series.

Arguments

A::AbstractArray: Time series.

fs::Real: Sampling rate of time series A.

freqmin::Real: Pass band low corner frequency.

freqmax::Real: Pass band high corner frequency.

pad::Int: Number of tapering points outside whitening band.

source

Noise.process_fft — Function.

process_fft(A::AbstractArray,freqmin::Float64,freqmax::Float64,fs::Float64;
            time_norm=false,to_whiten=false,corners=corners,
            zerophase=zerophase)

Arguments

A::AbstractArray: Array with time domain data.

fs::Float64: Sampling rate of data in A.

freqmin::Float64: minimum frequency for whitening.

freqmax::Float64: maximum frequency for whitening.

time_norm::Union{Bool,String}: time domain normalization to perform.

to_whiten::Bool: Apply whitening in frequency domain.

corners::Int: Number of corners in Butterworth filter.

zerophase::Bool: If true, apply Butterworth filter twice for zero phase change in output signal.

source

Noise.save_fft — Function.

save_fft(F::FFTData, OUT::String)

Save FFTData F to JLD2.

source