Filters

Filters - methods for filtering SeisData and SeisChannel objects.

Noise.bandpass! — Function.

bandpass!(A,freqmin,freqmax,fs,corners=4,zerophase=false)

Butterworth-Bandpass Filter.

Filter data A from freqmin to freqmax using corners corners.

Arguments

A::AbstractArray: Data to filter

freqmin::Float64: Pass band low corner frequency.

freqmax::Float64: Pass band high corner frequency.

fs::Float64: Sampling rate in Hz.

fs::Int: Filter corners / order.

zerophase::Bool: If True, apply filter once forwards and once backwards.

This results in twice the filter order but zero phase shift in the resulting filtered trace.

source

bandpass!(C,freqmin,freqmax,fs,corners=4,zerophase=false)

Butterworth-Bandpass Filter.

Filter data in C from freqmin to freqmax using corners corners.

Arguments

C::SeisChannel: SeisChannel to filter.

freqmin::Float64: Pass band low corner frequency.

freqmax::Float64: Pass band high corner frequency.

fs::Float64: Sampling rate in Hz.

fs::Int: Filter corners / order.

zerophase::Bool: If True, apply filter once forwards and once backwards.

This results in twice the filter order but zero phase shift in the resulting filtered trace.

source

bandpass!(S,freqmin,freqmax,fs,corners=4,zerophase=false)

Butterworth-Bandpass Filter.

Filter channels in S from freqmin to freqmax using corners corners.

Arguments

S::SeisData: SeisData to filter.

freqmin::Float64: Pass band low corner frequency.

freqmax::Float64: Pass band high corner frequency.

fs::Float64: Sampling rate in Hz.

fs::Int: Filter corners / order.

zerophase::Bool: If True, apply filter once forwards and once backwards.

This results in twice the filter order but zero phase shift in the resulting filtered trace.

source

Noise.bandstop! — Function.

bandstop!(A,freqmin,freqmax,fs,corners=4,zerophase=false)

Butterworth-Bandstop Filter.

Filter data A removing data between frequencies freqmin to freqmax using corners corners.

Arguments

A::AbstractArray: Data to filter

freqmin::Float64: Stop band low corner frequency.

freqmax::Float64: Stop band high corner frequency.

fs::Float64: Sampling rate in Hz.

fs::Int: Filter corners / order.

zerophase::Bool: If True, apply filter once forwards and once backwards.

This results in twice the filter order but zero phase shift in the resulting filtered trace.

source

bandstop!(C,freqmin,freqmax,fs,corners=4,zerophase=false)

Butterworth-Bandstop Filter.

Filter data in C removing data between frequencies freqmin to freqmax using corners corners.

Arguments

C::SeisChannel: SeisChannel to filter.

freqmin::Float64: Stop band low corner frequency.

freqmax::Float64: Stop band high corner frequency.

fs::Float64: Sampling rate in Hz.

fs::Int: Filter corners / order.

zerophase::Bool: If True, apply filter once forwards and once backwards.

This results in twice the filter order but zero phase shift in the resulting filtered trace.

source

Noise.lowpass! — Function.

lowpass(A,freq,fs,corners=4,zerophase=false)

Butterworth-Lowpass Filter.

Filter data A over certain frequency freq using corners corners.

Arguments

A::AbstractArray: Data to filter

freq::Float64: Filter corner frequency.

fs::Float64: Sampling rate in Hz.

fs::Int: Filter corners / order.

zerophase::Bool: If True, apply filter once forwards and once backwards.

This results in twice the filter order but zero phase shift in the resulting filtered trace.

source

lowpass(C,freq,fs,corners=4,zerophase=false)

Butterworth-Lowpass Filter.

Filter data in C over certain frequency freq using corners corners.

Arguments

C::SeisChannel: SeisChannel to filter.

freq::Float64: Filter corner frequency.

fs::Float64: Sampling rate in Hz.

fs::Int: Filter corners / order.

zerophase::Bool: If True, apply filter once forwards and once backwards.

This results in twice the filter order but zero phase shift in the resulting filtered trace.

source

lowpass(S,freq,fs,corners=4,zerophase=false)

Butterworth-Lowpass Filter.

Filter channels in S over certain frequency freq using corners corners.

Arguments

S::SeisData: SeisData to filter.

freq::Float64: Filter corner frequency.

fs::Float64: Sampling rate in Hz.

fs::Int: Filter corners / order.

zerophase::Bool: If True, apply filter once forwards and once backwards.

This results in twice the filter order but zero phase shift in the resulting filtered trace.

source

Noise.highpass! — Function.

highpass(A,freq,fs,corners=4,zerophase=false)

Butterworth-Highpass Filter.

Filter data A removing data below certain frequency freq using corners corners.

Arguments

A::AbstractArray: Data to filter

freq::Float64: Filter corner frequency.

fs::Float64: Sampling rate in Hz.

fs::Int: Filter corners / order.

zerophase::Bool: If True, apply filter once forwards and once backwards.

This results in twice the filter order but zero phase shift in the resulting filtered trace.

source

highpass(C,freq,fs,corners=4,zerophase=false)

Butterworth-Highpass Filter.

Filter data in C removing data below certain frequency freq using corners corners.

Arguments

C::SeisChannel: SeisChannel to filter.

freq::Float64: Filter corner frequency.

fs::Float64: Sampling rate in Hz.

fs::Int: Filter corners / order.

zerophase::Bool: If True, apply filter once forwards and once backwards.

This results in twice the filter order but zero phase shift in the resulting filtered trace.

source

highpass(S,freq,fs,corners=4,zerophase=false)

Butterworth-Highpass Filter.

Filter channels in S removing data below certain frequency freq using corners corners.

Arguments

S::SeisData: SeisData to filter.

freq::Float64: Filter corner frequency.

fs::Float64: Sampling rate in Hz.

fs::Int: Filter corners / order.

zerophase::Bool: If True, apply filter once forwards and once backwards.

This results in twice the filter order but zero phase shift in the resulting filtered trace.

source