Velocity Change

Post-Processing - methods for working with correlation data.

SeisNoise.mwcs — Function.

mwcs(ref, cur, fmin, fmax, fs, tmin, windowlength, windowstep, smoothinghalfwin)

Change in velocity measurement using the Moving Window Cross-Spectrum technique.

The current correlation cur is compared to the reference correlation ref. Both time series are sliced in several overlapping windows. Each slice is mean-adjusted and cosine-tapered (85% taper) before being Fourier- transformed to the frequency domain. $F_{cur}(ν)$ and $F_{ref}(ν)$ are the first halves of the Hermitian symmetric Fourier-transformed segments. The cross-spectrum $X(ν)$ is defined as $X(ν) = F_{ref}(ν) F_{cur}^*(ν)$ in which ${}^*$ denotes the complex conjugation. $X(ν)$ is then smoothed by convolution with a Hanning window. The similarity of the two time-series is assessed using the cross-coherency between energy densities in the frequency domain: $C(ν) = \frac{|\overline{X(ν))}|}{\sqrt{|\overline{F_{ref}(ν)|^2} |\overline{F_{cur}(ν)|^2}}}$ in which the over-line here represents the smoothing of the energy spectra for $F_{ref}$ and $F_{cur}$ and of the spectrum of $X$. The mean coherence for the segment is defined as the mean of $C(ν)$ in the frequency range of interest. The time-delay between the two cross correlations is found in the unwrapped phase, $ϕ(ν)$, of the cross spectrum and is linearly proportional to frequency: $ϕ_j = m. ν_j, m = 2 π δ t$ The time shift for each window between two signals is the slope $m$ of a weighted linear regression of the samples within the frequency band of interest. The weights are those introduced by [Clarke2011], which incorporate both the cross-spectral amplitude and cross-coherence, unlike [Poupinet1984]. The errors are estimated using the weights (thus the coherence) and the squared misfit to the modelled slope: $e_m = \sqrt{\sum_j{(\frac{w_j ν_j}{\sum_i{w_i ν_i^2}})^2}σ_{ϕ}^2}$ where $w$ are weights, $ν$ are cross-coherences and $σ_{ϕ}^2$ is the squared misfit of the data to the modelled slope and is calculated as $σ_{ϕ}^2 = \frac{\sum_j(ϕ_j - m ν_j)^2}{N-1}$ The output of this process is a table containing, for each moving window: the central time lag, the measured delay, its error and the mean coherence of the segment.

Arguments

ref::AbstractArray: Reference correlation.

cur::AbstractArray: Current correlation.

fmin:Float64: minimum frequency in the correlation [Hz]

fmax:Float64: maximum frequency in the correlation [Hz]

fs:Float64: Sampling rate of ref and cur [Hz]

tmin:Float64: The leftmost time lag [s]

window_len:Float64: The moving window length [s]

window_step:Float64: The step to jump for the moving window [s]

smoothing_half_win:Int: Defines the half length of the smoothing hanning window.

Returns

time_axis:Array{Float64,1}: Central time of each moving window [s]

dt:Array{Float64,1}: dt for each moving window

err:Array{Float64,1}: Errors for each moving window

mcoh:Array{Float64,1}: Mean coherence for each moving window

This code is a Julia translation of the Python code from MSNoise.

source

SeisNoise.mwcs_dvv — Function.

mwcs_dvv(time_axis,dt,err,coh,dtt_lag,dist,dtt_v,dtt_minlag,dtt_width,
         dtt_sides,min_coh, max_err, max_dt)

Regresses dv/v from dt/t measurements.

Solves dt = a + bt, where b = dv/v, a = instrumental drift, dt = time lag at time t. Solves with a weighted linear regression with weights equal to 1/error**2.

Arguments

time_axis:Array{Float64,1}: Central time of each moving window [s]

dt:Array{Float64,1}: dt for each moving window

err:Array{Float64,1}: Errors for each moving window

coh:Array{Float64,1}: Mean coherence for each moving window

dist:Float64: Distance between stations [km]

dtt_lag:String: Type of time lag 'dynamic' or 'static'. When dtt_lag is set to "dynamic", the inter-station distance is used to determine the minimum time lag

dtt_v:Float64: Velocity for minumum time lag. The velocity is determined by the user so that the minlag doesn't include the ballistic waves. If ballistic waves are visible with a velocity of 2 km/s, one could configure dtt_v=1.0. If stations are located 15 km apart, the minimum lag time will be set to 15 s.

minlag:Float64: Statically set minimum lag [s]

dtt_width:Float64: Width of the lag window used [s] i.e., if dttwidth = 30, and minlag = 15, window = 15s - 45s

dtt_sides:String: Sides of cross-correlation function to use. Either 'Both' or 'left'.

min_coh:Float64: Minimum allowed coherency between reference and current correlation

max_err:Float64: Maximum allowed error in dt/t regression

max_dt:Float64: Maximum allowed dt from MWCS [s]

Returns

m::Float64: dt/t for current correlation

em::Float64: Error for calulatoin of m

a::Float64: Intercept for regression calculation

ea::Float64: Error on intercept

m0::Float64: dt/t for current correlation with no intercept

em0::Float64: Error for calulatoin of m0

source

SeisNoise.stretching — Function.

stretching(ref,cur,t,window,fmin,fmax;dvmin,dvmax,ntrials)

This function compares the Reference waveform to stretched/compressed current waveforms to get the relative seismic velocity variation (and associated error). It also computes the correlation coefficient between the Reference waveform and the current waveform.

Arguments

ref::AbstractArray: Reference correlation.

cur::AbstractArray: Current correlation.

t::AbstractArray: time vector, common to both ref and cur.

window::AbstractArray: vector of the indices of the cur and ref windows on which you want to do the measurements

fmin::Float64: minimum frequency in the correlation [Hz]

fmax::Float64: maximum frequency in the correlation [Hz]

dvmin::Float64: minimum bound for the velocity variation; e.g. dvmin=-0.03 for -3% of relative velocity change

dvmax::Float64: maximum bound for the velocity variation; e.g. dvmin=0.03 for 3% of relative velocity change

ntrial::Int: number of stretching coefficient between dvmin and dvmax, no need to be higher than 100

Returns

dv::AFloat64: Relative Velocity Change dv/v (in %)

cc::Float64: Correlation coefficient between the reference waveform and the best stretched/compressed current waveform

cdp::Float64: Correlation coefficient between the reference waveform and the initial current waveform

ϵ::Array{Float64,1}: Vector of Epsilon values (ϵ =-dt/t = dv/v)

err::Float64: Errors in the dv/v measurements based on Weaver et al., 2011

allC::Array{Float64,1}: Vector of the correlation coefficient between the reference waveform and every stretched/compressed current waveforms

This code is a Julia translation of the Python code from Viens et al., 2018.

source

stretching(C,t,window,fmin,fmax;dvmin,dvmax,ntrials)

This function compares the Reference waveform to stretched/compressed current waveforms to get the relative seismic velocity variation (and associated error) for all correlations in CorrData C. It also computes the correlation coefficient between the Reference waveform and the current waveform.

Arguments

C::CorrData: Correlation for dv/v.

cur::AbstractArray: Current correlation.

t::AbstractArray: time vector, common to both ref and cur.

window::AbstractArray: vector of the indices of the cur and ref windows on which you want to do the measurements

fmin::Float64: minimum frequency in the correlation [Hz]

fmax::Float64: maximum frequency in the correlation [Hz]

dvmin::Float64: minimum bound for the velocity variation; e.g. dvmin=-0.03 for -3% of relative velocity change

dvmax::Float64: maximum bound for the velocity variation; e.g. dvmin=0.03 for 3% of relative velocity change

ntrial::Int: number of stretching coefficient between dvmin and dvmax, no need to be higher than 100

Returns

dv::Array{Float64,1}: Relative Velocity Change dv/v (in %)

cc::Array{Float64,1}: Correlation coefficient between the reference waveform and the best stretched/compressed current waveform

cdp::Array{Float64,1}: Correlation coefficient between the reference waveform and the initial current waveform

ϵ::Array{Float64,1}: Vector of Epsilon values (ϵ =-dt/t = dv/v)

err::Array{Float64,1}: Errors in the dv/v measurements based on Weaver et al., 2011

allC::Array{Float64,2}: Matrix of the correlation coefficient between the reference waveform and every stretched/compressed current waveforms

source