Pre-Processing

The pre-processing functions get raw data ready for correlation. Many of these rely on SeisIO's processing. A typical workflow includes the following steps:

Loading data

Filling gaps

Downsampling

Slicing day-long traces into smaller windows

Demeaning, detrending and tapering

Time-domain normalization

Here is an example workflow:

Load or download data

Data can be downloaded using SeisIO's get_data function.

julia> using SeisIO, SeisNoise
julia> S = get_data("IRIS","TA.V04C..BHZ",s="2006-02-01",t="2006-02-02")
SeisData with 1 channels (1 shown)
    ID: TA.V04C..BHZ                       
  NAME: TA.V04C..BHZ                       
   LOC: 0.0 N, 0.0 E, 0.0 m                
    FS: 40.0                               
  GAIN: 1.0                                
  RESP: c = 0.0, 0 zeros, 0 poles          
 UNITS:                                    
   SRC:                                    
  MISC: 0 entries                          
 NOTES: 0 entries                          
     T: 2006-02-01T00:00:00.020 (0 gaps)   
     X: +1.044e-06                         
        +1.060e-06                         
            ...                            
        +6.282e-07                         
        (nx = 3456000)                     
     C: 0 open, 0 total
julia> writesac(S)

Data can be read from disk using the read_data function .

julia> S = read_data("sac","2006.032.00.00.00.002.TA.V04C..BHZ.R.SAC")
SeisData with 1 channels (1 shown)
    ID: TA.V04C..BHZ                       
  NAME:                                    
   LOC: 0.0 N, 0.0 E, 0.0 m                
    FS: 40.0                               
  GAIN: 1.0                                
  RESP: c = 0.0, 0 zeros, 0 poles          
 UNITS:                                    
   SRC: 2006.032.00.00.00.002.TA.V04C..BH…
  MISC: 0 entries                          
 NOTES: 1 entries                          
     T: 2006-02-01T00:00:00.002 (0 gaps)   
     X: +1.044e-06                         
        +1.060e-06                         
            ...                            
        +6.282e-07                         
        (nx = 3456000)                     
     C: 0 open, 0 total

Merge and ungap data

Use the merge! method to merge S to have one channel. The ungap method replaces gaps in S[1].x with the mean of S[1].x.

julia> merge!(S)
julia> ungap!(S)

Downsample Data

Downsample S to 20 Hz using process_raw!. process_raw!:

Removes mean from each channel in S.

Detrends each channel in S.

Downsamples data to sampling rate fs

Phase-shifts data to begin at 00:00:00.0

This is important for data that does not begin exactly on the sampling rate, e.g. the starttime is 2006-02-01T00:00:00.020.

julia> process_raw!(S,20.)
julia> S
SeisData with 1 channels (1 shown)
    ID: TA.V04C..BHZ                       
  NAME: TA.V04C..BHZ                       
   LOC: 0.0 N, 0.0 E, 0.0 m                
    FS: 20.0                               
  GAIN: 1.0                                
  RESP: c = 0.0, 0 zeros, 0 poles          
 UNITS:                                    
   SRC:                                    
  MISC: 0 entries                          
 NOTES: 1 entries                          
     T: 2006-02-01T00:00:00.000 (0 gaps)   
     X: +9.289e-10                         
        -9.248e-10                         
            ...                            
        -9.287e-10                         
        (nx = 1728000)                     
     C: 0 open, 0 total

Note that now S has sampling rate fs = 20.0, has half as many points (nx = 1728001) as before and the starttime has changed to 2006-02-01T00:00:00.000.

Detrending and Demeaning

The demean and detrend functions are applied column wise.

julia> A = reshape(collect(1.:10.),5,2)
5×2 Array{Float64,2}:
 1.0   6.0
 2.0   7.0
 3.0   8.0
 4.0   9.0
 5.0  10.0
julia> demean(A)
5×2 Array{Float64,2}:
 -2.0  -2.0
 -1.0  -1.0
  0.0   0.0
  1.0   1.0
  2.0   2.0
julia> detrend(A)
5×2 Array{Float64,2}:
6.66134e-16  3.55271e-15
1.33227e-15  5.32907e-15
2.22045e-15  7.10543e-15
2.66454e-15  7.10543e-15
3.55271e-15  1.06581e-14

demean and detrend also work on RawData and CorrData structure. If R is a RawData, then the windowed data stored in R.x can be detrended in-place using

detrend!(R)

or allocated to a new RawData using

Rd = detrend(R)

Amplitude Normaliztion

Time-domain normalization is used to suppress high-amplitude signals, such as earthquakes or instrumental irregularities. SeisNoise provides time-normalization functions for:

one-bit normalization: onebit

clipping: clip

running mean normalization: smooth

suppressing high-amplitude signals: mute

Filtering

SeisNoise.jl provides bandpass, bandstop, lowpass and highpass filters built from DSP.jl. Due to multiple dispatch in Julia, filter functions in SeisNoise.jl work on either the data in RawData or CorrData objects or directly with Julia Arrays. SeisNoise.jl uses Butterworth filters with a default of 4 corners.

Filtering a RawData or CorrData

julia> freqmin, freqmax = 1., 10. # low and high frequencies in Hz
julia> corners = 4 # number of corners in Butterworth filter
julia> zerophase = true # if true, apply filter twice for no phase shifting
julia> S = get_data("IRIS","TA.V04C..BHZ",s="2006-02-01T00:00:00",t="2006-02-01T01:00:00")
julia> SeisIO.demean!(S) # remove mean
julia> SeisIO.detrend!(S) # remove linear trend
julia> SeisIO.taper!(S) # taper - defaults to 5% taper on either side of the trace
julia> bandpass(S,freqmin,freqmax,corners=corners,zerophase=zerophase)

Nyquist Frequency

Filtering above the Nyquist frequency will give a warning, if using a lowpass or bandpass filter, while using a highpass above the Nyquist frequency will throw an error.

julia> S[1].fs / 2. # Nyquist frequency is half sampling rate
20.
julia> freqmax = S[1].fs / 2. + 5.
25.  
julia> bandpass(S,freqmin,freqmax,corners=corners,zerophase=zerophase)
┌ Warning: Selected high corner frequency (25.0) of bandpass is at or
│        above Nyquist (20.0). Applying a high-pass instead.
└
julia> lowpass(S,freqmax,corners=corners,zerophase=zerophase)
┌ Warning: Selected corner frequency (25.0) is
│ above Nyquist (20.0). Setting Nyquist as high corner.
└
julia> highpass(S,freqmax,corners=corners,zerophase=zerophase)
ERROR: frequencies must be less than the Nyquist frequency 20.0

Pre-Processing Functions

SeisIO.demean! — Method

demean!(A::AbstractArray{<:AbstractFloat})

Remove mean from array A.

SeisIO.detrend! — Method

detrend!(X::AbstractArray{<:AbstractFloat})

Remove linear trend from array X using least-squares regression.

SeisIO.taper! — Method

taper!(A,fs; maxpercentage=0.05, maxlength=20.)

Taper a time series A with samplingrate fs. Defaults to 'hann' window. Uses smallest of `maxpercentage*fsormax_length`.

Arguments

A::AbstractArray: Time series.

fs::AbstractFloat: Sampling rate of time series A.

max_percentage::float: Decimal percentage of taper at one end (ranging from 0. to 0.5).

max_length::Real: Length of taper at one end in seconds.

SeisNoise.hanningwindow — Method

hanningwindow(A,n)

Generate hanning window of length n.

Hanning window is sin(n*pi)^2.

SeisIO.resample! — Method

resample!(C, fs)

Resample data in CorrDataC to new sampling ratefs`.

Arguments

C::CorrData: CorrData to resample.

fs::Real: New sampling frequency.

SeisIO.resample! — Method

resample!(R, fs)

Resample data in RawData R to new sampling rate fs.

Arguments

R::RawData: RawData to resample.

fs::Real: New sampling frequency.

SeisNoise.bandpass! — Method

bandpass!(A,freqmin,freqmax,fs,corners=4,zerophase=true)

Butterworth-Bandpass Filter.

Filter data A from freqmin to freqmax using corners corners.

Arguments

A::AbstractArray: Data to filter

freqmin::Real: Pass band low corner frequency.

freqmax::Real: Pass band high corner frequency.

fs::Real: Sampling rate in Hz.

fs::Int: Filter corners / order.

zerophase::Bool: If True, apply filter once forwards and once backwards.

This results in twice the filter order but zero phase shift in the resulting filtered trace.

SeisNoise.bandstop! — Method

bandstop!(A,freqmin,freqmax,fs,corners=4,zerophase=true)

Butterworth-Bandstop Filter.

Filter data A removing data between frequencies freqmin to freqmax using corners corners.

Arguments

A::AbstractArray: Data to filter

freqmin::Real: Stop band low corner frequency.

freqmax::Real: Stop band high corner frequency.

fs::Real: Sampling rate in Hz.

fs::Int: Filter corners / order.

zerophase::Bool: If True, apply filter once forwards and once backwards.

This results in twice the filter order but zero phase shift in the resulting filtered trace.

SeisNoise.envelope — Method

envelope(A)

Envelope of a function.

Computes the upper and lower envelopes of the given function.

Arguments

A::AbstractArray: Data to make envelope of.

SeisNoise.gpufilter! — Method

gpufilter!(A,responsetype,designmethod)

Apply filter to array A on the GPU.

Arguments

A::AbstractGPUArray: Array on the GPU to filter

responsetype::FilterType: DSP.jl filter representation

designmethod::ZeroPoleGain: Filter representation in terms of zeros z, poles p, and

gain k.

SeisNoise.highpass! — Method

highpass(A,freq,fs,corners=4,zerophase=true)

Butterworth-Highpass Filter.

Filter data A removing data below certain frequency freq using corners corners.

Arguments

A::AbstractArray: Data to filter

freq::Real: Filter corner frequency.

fs::Real: Sampling rate in Hz.

fs::Int: Filter corners / order.

zerophase::Bool: If True, apply filter once forwards and once backwards.

This results in twice the filter order but zero phase shift in the resulting filtered trace.

SeisNoise.lowpass! — Method

lowpass(A,freq,fs,corners=4,zerophase=true)

Butterworth-Lowpass Filter.

Filter data A over certain frequency freq using corners corners.

Arguments

A::AbstractArray: Data to filter

freq::Real: Filter corner frequency.

fs::Real: Sampling rate in Hz.

fs::Int: Filter corners / order.

zerophase::Bool: If True, apply filter once forwards and once backwards.

This results in twice the filter order but zero phase shift in the resulting filtered trace.

SeisNoise.process_raw! — Function

process_raw!(S,fs)

Pre-process raw seismic data.

Removes mean from each channel in S.

Detrends each channel in S.

Downsamples data to sampling rate fs

Phase-shifts data to begin at 00:00:00.0

Arguments

S::SeisData: SeisData structure.

fs::Real: Sampling rate to downsample S.

process_raw!(C,fs)

Pre-process raw SeisChannel.

Removes mean from data in C.

Detrends data in C.

Downsamples data to sampling rate fs

Phase-shifts data to begin at 00:00:00.0

Arguments

C::SeisChannel: SeisChannel structure.

fs::Real: Sampling rate to downsample C.

SeisNoise.onebit! — Function

onebit!(R)

One-bit amplitude modification of RawData R.

SeisNoise.clip! — Function

clip(A,factor)

Truncate array A at factor times the root mean square of each column.

#Arguments

A::AbstractArray: N-d time series array

factor::Real:

f::Function: Input statistical function (e.g. rms, var, std, mad)

dims: Dimension of A to apply clipping (defaults to 1)

Missing docstring.

Missing docstring for clamp!. Check Documenter's build log for details.

SeisNoise.mute! — Function

mute(A,factor)

Set high amplitudes in array A to zero. Uses median of envelope of A to find outliers.