CorrData

CorrData - Objects for ambient noise cross-correlations.

SeisNoise uses a custom structure CorrData for holding cross-correlations. To create an empty CorrData object, use the CorrData() function. CorrData are created with the compute_cc function.

julia> using SeisNoise
julia> CorrData()
CorrData with 0 Corrs
      NAME: …
        ID: …
       LOC: 0.0 N, 0.0 E, 0.0 m
      COMP: …
   ROTATED: …
 CORR_TYPE: …
        FS: 0.0
      GAIN: 1.0
   FREQMIN: 0.0
   FREQMAX: 0.0
    CC_LEN: …
   CC_STEP: …
  WHITENED: …
 TIME_NORM: …
      RESP: c = 0.0, 0 zeros, 0 poles
      MISC: …
     NOTES: …
    MAXLAG: 0.0
         T: …
      CORR: …

A non-empty CorrData looks like this:

julia> C
CorrData with 188 Corrs
      NAME: "TA.V04C..BHZ.TA.V05C..BHZ"        
        ID: "2006-02-01"                       
       LOC: 0.0 N, 0.0 E, 0.0 m
      COMP: "ZZ"                               
   ROTATED: false                              
 CORR_TYPE: "coherence"                        
        FS: 40.0
      GAIN: 1.0
   FREQMIN: 0.1
   FREQMAX: 0.2
    CC_LEN: 1800                               
   CC_STEP: 450                                
  WHITENED: false                              
 TIME_NORM: false                              
      RESP: c = 0.0, 0 zeros, 0 poles
      MISC: 0 entries                          
     NOTES: 2 entries                          
    MAXLAG: 80.0
         T: 2006-02-01T00:07:30.000            …
      CORR: 6401×188 Array{Float32,2}     

To access the correlations stored in C just do C.corr

julia> C.corr
6401×188 Array{Float32,2}:
 3.86871e-5   -0.00024922    0.000610002   0.000504283  …  -0.000187752   0.000239163   9.5009e-5  
 3.8662e-5    -0.000238532   0.000607226   0.000493771     -0.000165946   0.000240476   7.08343e-5
 3.81959e-5   -0.000227481   0.000602314   0.000481894     -0.000144085   0.000240927   4.63148e-5
 3.76619e-5   -0.000216617   0.000596438   0.00046825      -0.000122814   0.000241393   2.16217e-5
 3.6887e-5    -0.000205915   0.000588669   0.00045305      -0.000101321   0.000241209  -3.33542e-6
 3.58405e-5   -0.000195269   0.000579826   0.000437172  …  -8.06184e-5    0.000240315  -2.85994e-5
 3.4513e-5    -0.000184763   0.000569293   0.000419673     -5.94726e-5    0.00023961   -5.38829e-5
 3.32238e-5   -0.000174677   0.000557559   0.000400775     -3.93204e-5    0.00023862   -7.91602e-5
 3.15859e-5   -0.000164477   0.000543592   0.000381227     -1.84086e-5    0.000236609  -0.000104519
 3.02724e-5   -0.00015467    0.000529009   0.000360099      1.71954e-6    0.000234192  -0.000129835
 2.85284e-5   -0.00014533    0.000512517   0.000337998  …   2.13676e-5    0.000230932  -0.000155009
 2.6588e-5    -0.000135612   0.000495403   0.000314866      4.03696e-5    0.000228341  -0.000180198
 2.4749e-5    -0.000126298   0.000476726   0.000290888      5.9316e-5     0.000224295  -0.000205199
 ⋮                                                      ⋱   ⋮                                      
 8.30166e-5    0.000492762   1.49007e-5   -0.000359526      0.000157452  -0.000232154  -0.000343711
 8.6317e-5     0.00049216    5.2199e-6    -0.000358585      0.000152415  -0.000246062  -0.000354792
 8.98544e-5    0.000490826  -4.95815e-6   -0.000357571  …   0.000146579  -0.000261209  -0.000365518
 9.39397e-5    0.000489367  -1.49089e-5   -0.00035631       0.000140762  -0.000275435  -0.000375918
 9.81829e-5    0.000486619  -2.55125e-5   -0.000354263      0.000135029  -0.000290028  -0.000385996
 0.000102882   0.000483545  -3.53862e-5   -0.000352179      0.000128825  -0.000304768  -0.000395605
 0.000107735   0.000480358  -4.58485e-5   -0.000349237      0.000122449  -0.000318608  -0.000404866
 0.000112649   0.000476104  -5.58984e-5   -0.000346363  …   0.000116377  -0.000331822  -0.000413702
 0.000118011   0.000471311  -6.60669e-5   -0.000342526      0.000110101  -0.000346527  -0.000422143
 0.000123389   0.000465821  -7.64809e-5   -0.000338501      0.000103648  -0.000360307  -0.000430103
 0.000128818   0.000459877  -8.66791e-5   -0.000334324      9.68363e-5   -0.000374563  -0.000437496
 0.000134595   0.00045348   -9.59272e-5   -0.000329577      9.07089e-5   -0.000386802  -0.000444421
 0.00014004    0.000446596  -0.000105674  -0.00032479   …   8.39108e-5   -0.000400697  -0.000450801

Note

By convention in Julia, data are stored column-wise. Here, C contains 188 individual correlations.

SeisNoise.CorrData — Type.

CorrData

A structure for cross-correlations of ambient noise data.

Fields: CorrData

Field	Description

:name	Freeform channel name in NET1.STA1.LOC1.CHAN1.NET2.STA2.LOC2.CHAN2 form

:id	Channel ids. use NET.STA.LOC.CHAN format when possible.

:loc	Location (position) object for station 1.

:fs	Sampling frequency in Hz.

:gain	Scalar gain; divide data by the gain to convert to units

:freqmin	Minimum frequency for whitening.

:freqmax	Maximum frequency for whitening.

:cc_len	Length of each correlation in seconds.

:cc_step	Spacing between correlation windows in seconds.

:whitened	Whitening applied.

:time_norm	Apply one-bit whitening with "one_bit".

:resp	Instrument response object, format [zeros poles]

:misc	Dictionary for non-critical information.

:notes	Timestamped notes; includes automatically-logged acquisition and

processing information.

:dist	Distance between station 1 and station 2 in Km.

:azi	Azimuth from station 1 and station 2 in degrees.

:baz	Back azimuth between station 1 and station 2 in degrees.

:maxlag	Maximum lag time in seconds to keep from correlations.

:t	Starttime of each correlation.

:corr	Correlations stored in columns.

source