Correlation

Computing Correlations - methods for computing correlations from FFTs.

SeisNoise.clean_up! — Function.

clean_up!(A,freqmin,freqmax,fs)

Demean, detrend, taper and filter time series.

Arguments

A::AbstractArray: Time series.

fs::Float64: Sampling rate of time series A in Hz.

freqmin::Float64: Pass band low corner frequency in Hz.

freqmax::Float64: Pass band high corner frequency in Hz.

source

SeisNoise.correlate — Function.

correlate(FFT1, FFT2, N, maxlag, smoothing_half_win=20,
          corr_type='cross-correlation')

Cross-correlate ambient noise data in the frequency domain.

Cross-correlation can be done using one of three options:

Cross-correlation: $C_{AB}(ω) = u_A(ω) u^∗_B(ω)$

Coherence: $C_{AB}(ω) = rac{u_A(ω) u^∗_B(ω)}{|u_A(ω)| |u^_B(ω)|}$

Deconvolution: $C_{AB}(ω) = rac{u_A(ω) u^∗_B(ω)}{|u^_B(ω)|^2}$

Arguments

FFT1::AbstractArray: Complex Array of fourier transform of ambient noise data.

FFT2::AbstractArray: Complex Array of fourier transform of ambient noise data.

N::Int: Number of input data, equal to cc_len * fs.

maxlag::Float64: Maximum lag time (in seconds) in cross-correlation to save, e.g. maxlag = 20. will save lag times = -20.:20. s.

smoothing_half_win::Int: Number of points to smooth spectrum of cross-correlations if using deconvolution or coherence.

corr_type::String: Type of correlation: cross-correlation, coherence or deconv.

source

SeisNoise.compute_cc — Function.

compute_cc(FFT1::FFTData, FFT2::FFTData, maxlag::Float64;
           smoothing_half_win::Int=20,
           corr_type::String="cross-correlation")

Cross-correlate ambient noise data in the frequency domain.

Cross-correlation can be done using one of three options:

Cross-correlation: $C_{AB}(ω) = u_A(ω) u^∗_B(ω)$

Coherence: $C_{AB}(ω) = rac{u_A(ω) u^∗_B(ω)}{|u_A(ω)| |u^_B(ω)|}$

Deconvolution: $C_{AB}(ω) = rac{u_A(ω) u^∗_B(ω)}{|u^_B(ω)|^2}$

Arguments

fft1::FFTData: FFTData object of fft'd ambient noise data.

fft2::FFTData: FFTData object of fft'd ambient noise data.

maxlag::Float64: Maximum lag time (in seconds) in cross-correlation to save, e.g. maxlag = 20. will save lag times = -20.:20. s.

smoothing_half_win::Int: Number of points to smooth spectrum of cross-correlations if using deconvolution or coherence.

corr_type::String: Type of correlation: cross-correlation, coherence or deconv.

source

SeisNoise.save_corr — Function.

save_corr(C::CorrData, OUT::String)

Save CorrData C to JLD2.

source

SeisNoise.load_corr — Function.

load_corr(filename,chan,day=day)

Loads CorrData for channel chan on day day from JLD2 file filename.

source

SeisNoise.stack! — Function.

stack!(C)

Stack correlation by time interval. The default is to stack by day. Using allstack == true will stack all available correlations. To use phase-weighted stack, specify the amount of phase_smoothing in seconds.

Arguments

C::CorrData: Correlation data.

interval::Union{Month,Day,Hour,Second}: Interval over which to stack C.

allstack::Bool: If true, stack all data.

phase_smoothing::Float64: Enables phase-weighted stacking. phase_smoothing is the time window in seconds for phase smoothing in the phase-weighted stack.

source