Pre-Processing

Pre-Processing - methods for cleaning raw noise data.

Noise.process_raw! — Function.

process_raw!(S,fs)

Pre-process raw seismic data.

Checks:

sample rate is fs

downsamples data

checks for gaps in data

phase-shifts data to begin at 00:00:00.0

Arguments

S::SeisChannel: SeisData structure.

fs::Float64: Sampling rate to downsample S.

source

Noise.downsample — Function.

downsample(C::SeisChannel,fs::Real)

Downsample SeisChannel sampling rate to frequency fs.

For best results, lowpass filter data to fs before downsampling. Implements the weighted average slopes interpolation scheme proposed in [Wiggins1976] for evenly sampled data from obspy.signal.interpolation.weightedaverageslopes.

source

downsample(S::SeisData, fs::Float64)

Downsample SeisData sampling rate to frequency fs.

source

Noise.slide — Function.

slide(C::SeisChannel, window_length::Real, cc_step::Real)

Generate equal length sliding windows into an array.

source

Noise.start_end — Function.

start_end(C::SeisChannel)

Return start and endtimes of SeisChannel in DateTime format.

source

start_end(S::SeisData)

Return start and endtimes of SeisChannel in DateTime format.

source

Noise.nearest_start_end — Function.

nearest_start_end(C::SeisChannel, cc_len::Int, cc_step::Int)

Return best possible start, end times for data in C given the c

source

nearest_start_end(D::DateTime, cc_len::Int, cc_step::Int)

Return best possible start, end times for given starttime D

source